Change CreateTeam component and move styles to the assets template folder
