+++ src/assets/styles/templates/CreateTeam.js
@@ -1,4 +1,4 @@
-import BodyContainer from "../../atoms/BodyContainer";
+import BodyContainer from "../atoms/BodyContainer";
 import styled from "styled-components";
 
 export const BodyRow = styled(BodyContainer)`
@@ -21,7 +21,7 @@ export const Form = styled.form`
   background-color: white;
   width: 50%;
   height: 35vh;
-  border: 1px solid lightgray;
+  border: 1px solid ${props => props.theme.color.grey.border};
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -38,16 +38,11 @@ export const Form = styled.form`
     width: 50%;
     padding: 10px;
     margin: 10px;
-    border: 1px solid lightgray;
+    border: 1px solid ${props => props.theme.color.grey.border};
     border-radius: 6px;
   }
   button {
     width: 50%;
     margin: 10px;
   }
-`;
-
-export const Title = styled.span`
-    color: "#273F92";
-    background-color: "aliceblue"
 `;
